debugging

+++ frontend/src/components/WorkoutDetails.js
@@ -1,12 +1,22 @@
-import {formatDistanceToNow} from "date-fns/formatDistanceToNow"
 import useWorkoutContext from "../hooks/useWorkoutContext"
+import useAuthContext from "../hooks/useAuthContext"
+import {formatDistanceToNow} from "date-fns/formatDistanceToNow"
 
 const WorkoutDetails = ({w})=>{
   const {dispatch} = useWorkoutContext()
+  const {user} = useAuthContext()
 
   const handleClick = async()=>{
+    if(!user){
+      return
+    }
+
     const response = await fetch("/api/workouts/" + w._id,{
-      method:"DELETE"
+      method:"DELETE",
+      headers:{
+        "Content-Type":"application/json",
+        "Authorization": `Bearer ${user.token}`
+      }
     })
 
     const json = await response.json()
